requirements.txt
setuptools
path.py
pygame
entrypoint2

